Реалізуав додавання документів і пофіксив деякі помилки

+++ Core/Entities/PalletType.cs
@@ -11,6 +11,6 @@ namespace Domain.Entities
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public int RequiredCapacity { get; set; }
+        public double RequiredCapacity { get; set; }
     }
 }
